Refactor: move editors ui to a folder
