+++ pinry-spa/src/components/editors/PinEditorUI.vue
@@ -27,7 +27,7 @@
 </template>
 
 <script>
-import API from './api';
+import API from '../api';
 
 export default {
   name: 'Editor',
